Blockfrost.Types.Cardano.Accounts

Description

Responses for Cardano accounts queries

Documentation

data AccountInfo

Information about an account, identified by its stake address

Constructors

AccountInfo

Fields

_accountInfoStakeAddress :: Address

Bech32 stake address

_accountInfoActive :: Bool

Registration state of an account

_accountInfoActiveEpoch :: Maybe Integer

Epoch of the most recent action - registration or deregistration

_accountInfoControlledAmount :: Lovelaces

Balance of the account in Lovelaces

_accountInfoRewardsSum :: Lovelaces

Sum of all funds rewards for the account in the Lovelaces

_accountInfoWithdrawalsSum :: Lovelaces

Sum of all the withdrawals for the account in the Lovelaces

_accountInfoReservesSum :: Lovelaces

Sum of all funds from reserves for the account in the Lovelaces

_accountInfoTreasurySum :: Lovelaces

Sum of all funds from treasury for the account in the Lovelaces

_accountInfoWithdrawableAmount :: Lovelaces

Sum of available rewards that haven't been withdrawn yet for the account in the Lovelaces

_accountInfoPoolId :: Maybe PoolId

Bech32 pool ID that owns the account

data AccountReward

Reward received by an account

Constructors

AccountReward

Fields

_accountRewardEpoch :: Epoch

Epoch of the associated reward

_accountRewardAmount :: Lovelaces

Rewards for given epoch in Lovelaces

_accountRewardPoolId :: PoolId

Bech32 pool ID being delegated to

data AccountHistory

History of accounts stake delegation

Constructors

AccountHistory

Fields

_accountHistoryActiveEpoch :: Integer

Epoch in which the stake was active

_accountHistoryAmount :: Lovelaces

Stake amount in Lovelaces

_accountHistoryPoolId :: PoolId

Bech32 ID of pool being delegated to

data AccountDelegation

Account delegations and associated transaction IDs

Constructors

AccountDelegation

Fields

_accountDelegationActiveEpoch :: Epoch

Epoch in which the delegation becomes active

_accountDelegationTxHash :: TxHash

Hash of the transaction containing the delegation

_accountDelegationAmount :: Lovelaces

Rewards for given epoch in Lovelaces

_accountDelegationPoolId :: PoolId

Bech32 ID of pool being delegated to

data AccountRegistration

Account (de)registration

Constructors

AccountRegistration

Fields

_accountRegistrationAction :: AccountRegistrationAction

Action in the certificate

_accountRegistrationTxHash :: TxHash

Hash of the transaction containing the (de)registration certificate

data AccountRegistrationAction

Registration action

Constructors

Registered

Deregistered

data AccountWithdrawal

Withdrawal from an account

Constructors

AccountWithdrawal

Fields

_accountWithdrawalAmount :: Lovelaces

Withdrawal amount in Lovelaces

_accountWithdrawalTxHash :: TxHash

Hash of the transaction containing the withdrawal

data AccountMir

Account MIR (Move Instantaneous Reward)

Constructors

AccountMir

Fields

_accountMirAmount :: Lovelaces

MIR amount in Lovelaces

_accountMirTxHash :: TxHash

Hash of the transaction containing the MIR

newtype AddressAssociated

Address associated with an account address

Constructors

AddressAssociated

Fields

_addressAssociatedAddress :: Address