Blockfrost.Types.Cardano.Assets

Description

Responses for Cardano asset queries

Documentation

data AssetInfo

Asset information, result of listing assets

Constructors

AssetInfo

Fields

_assetInfoAsset :: Text

_assetInfoQuantity :: Quantity

data AssetDetails

Details of an asset

Constructors

AssetDetails

Fields

_assetDetailsAsset :: Text

Hex-encoded asset full name

_assetDetailsPolicyId :: PolicyId

Policy ID of the asset

_assetDetailsAssetName :: Maybe Text

Hex-encoded asset name of the asset

_assetDetailsFingerprint :: Text

CIP14 based user-facing fingerprint

_assetDetailsQuantity :: Quantity

Current asset quantity

_assetDetailsInitialMintTxHash :: TxHash

ID of the initial minting transaction

_assetDetailsMintOrBurnCount :: Integer

Count of mint and burn transactions

_assetDetailsOnchainMetadata :: Maybe AssetOnChainMetadata

On-chain metadata stored in the minting transaction under label 721, community discussion around the standard ongoing at https://github.com/cardano-foundation/CIPs/pull/85

_assetDetailsMetadata :: Maybe AssetMetadata

data AssetOnChainMetadata

On-chain metadata stored in the minting transaction under label 721, community discussion around the standard ongoing at https://github.com/cardano-foundation/CIPs/pull/85

Constructors

AssetOnChainMetadata

Fields

_assetOnChainMetadataName :: Text

Name of the asset

_assetOnChainMetadataImage :: Text

URI of the associated asset TODO: in schema has `additionalProperties: true` so it can carry more arbitrary properties, keep as Value?

data AssetMetadata

Asset metadata obtained from Cardano token registry https://github.com/cardano-foundation/cardano-token-registry

Constructors

AssetMetadata

Fields

_assetMetadataName :: Text

Asset name

_assetMetadataDescription :: Text

Asset description

_assetMetadataTicker :: Maybe Text

_assetMetadataUrl :: Maybe Text

Asset website

_assetMetadataLogo :: Maybe Text

Base64 encoded logo of the asset

_assetMetadataDecimals :: Maybe Int

Number of decimal places of the asset unit

data AssetHistory

History of an asset

Constructors

AssetHistory

Fields

_assetHistoryTxHash :: TxHash

Hash of the transaction containing the asset action

_assetHistoryAmount :: Quantity

Asset amount of the specific action

_assetHistoryAction :: AssetAction

Action executed upon the asset policy

data AssetAction

Action of the asset. Created (Minted) or destroyed (Burned).

Constructors

Minted

Burned

data AssetTransaction

Transaction of an asset

Constructors

AssetTransaction

Fields

_assetTransactionTxHash :: TxHash

Hash of the transaction

_assetTransactionTxIndex :: Integer

Transaction index within the block

_assetTransactionBlockHeight :: Integer

Block height

data AssetAddress

An address containing specific asset

Constructors

AssetAddress

Fields

_assetAddressAddress :: Address

Address containing the specific asset

_assetAddressQuantity :: Quantity

Asset quantity on the specific address