Blockfrost.Types.Cardano.Epochs

Description

Responses for Cardano epoch quries

Documentation

data EpochInfo

Information about an epoch

Constructors

EpochInfo

Fields

_epochInfoEpoch :: Epoch

Epoch number

_epochInfoStartTime :: POSIXTime

Unix time of the start of the epoch

_epochInfoEndTime :: POSIXTime

Unix time of the end of the epoch

_epochInfoFirstBlockTime :: POSIXTime

Unix time of the first block of the epoch

_epochInfoLastBlockTime :: POSIXTime

Unix time of the last block of the epoch

_epochInfoBlockCount :: Integer

Number of blocks within the epoch

_epochInfoTxCount :: Integer

Number of transactions within the epoch

_epochInfoOutput :: Lovelaces

Sum of all the transactions within the epoch in Lovelaces

_epochInfoFees :: Lovelaces

Sum of all the fees within the epoch in Lovelaces

_epochInfoActiveStake :: Maybe Lovelaces

Sum of all the active stakes within the epoch in Lovelaces

data PoolStakeDistribution

Stake distribution for an epoch for specific pool

Constructors

PoolStakeDistribution

Fields

_poolStakeDistributionStakeAddress :: Address

Stake address

_poolStakeDistributionAmount :: Lovelaces

Amount of active delegated stake in Lovelaces

data ProtocolParams

Protocol parameters

Constructors

ProtocolParams

Fields

_protocolParamsEpoch :: Epoch

Epoch number

_protocolParamsMinFeeA :: Integer

The linear factor for the minimum fee calculation for given epoch

_protocolParamsMinFeeB :: Integer

The constant factor for the minimum fee calculation

_protocolParamsMaxBlockSize :: Integer

Maximum block body size in Bytes

_protocolParamsMaxTxSize :: Integer

Maximum transaction size

_protocolParamsMaxBlockHeaderSize :: Integer

Maximum block header size

_protocolParamsKeyDeposit :: Lovelaces

The amount of a key registration deposit in Lovelaces

_protocolParamsPoolDeposit :: Lovelaces

The amount of a pool registration deposit in Lovelaces

_protocolParamsEMax :: Integer

Epoch bound on pool retirement

_protocolParamsNOpt :: Integer

Desired number of pools

_protocolParamsA0 :: Double

Pool pledge influence

_protocolParamsRho :: Double

Monetary expansion

_protocolParamsTau :: Double

Treasury expansion

_protocolParamsDecentralisationParam :: Double

Percentage of blocks produced by federated nodes

_protocolParamsExtraEntropy :: Maybe Value

Seed for extra entropy

_protocolParamsProtocolMajorVer :: Integer

Accepted protocol major version

_protocolParamsProtocolMinorVer :: Integer

Accepted protocol minor version

_protocolParamsMinUtxo :: Lovelaces

Minimum UTXO value

_protocolParamsMinPoolCost :: Lovelaces

Minimum stake cost forced on the pool

_protocolParamsNonce :: Text

Epoch number only used once cost models https://github.com/input-output-hk/cardano-db-sync/pull/758

_protocolParamsPriceMem :: Double

The per word cost of script memory usage

_protocolParamsPriceStep :: Double

The cost of script execution step usage

_protocolParamsMaxTxExMem :: Quantity

The maximum number of execution memory allowed to be used in a single transaction

_protocolParamsMaxTxExSteps :: Quantity

The maximum number of execution steps allowed to be used in a single transaction

_protocolParamsMaxBlockExMem :: Quantity

The maximum number of execution memory allowed to be used in a single block

_protocolParamsMaxBlockExSteps :: Quantity

The maximum number of execution steps allowed to be used in a single block

_protocolParamsMaxValSize :: Quantity

The maximum Val size

_protocolParamsCollateralPercent :: Integer

The percentage of the transactions fee which must be provided as collateral when including non-native scripts

_protocolParamsMaxCollateralInputs :: Integer

The maximum number of collateral inputs allowed in a transaction

_protocolParamsCoinsPerUtxoWord :: Lovelaces

The cost per UTxO word

data StakeDistribution

Active stake distribution for an epoch

Constructors

StakeDistribution

Fields

_stakeDistributionStakeAddress :: Address

Stake address

_stakeDistributionPoolId :: PoolId

Bech32 prefix of the pool delegated to

_stakeDistributionAmount :: Lovelaces

Amount of active delegated stake in Lovelaces