Blockfrost.Types.Cardano.Genesis

Description

Blockchain genesis

Documentation

data Genesis

Information about blockchain genesis

Constructors

Genesis

Fields

_genesisActiveSlotsCoefficient :: Double

The proportion of slots in which blocks should be issued

_genesisUpdateQuorum :: Integer

Determines the quorum needed for votes on the protocol parameter updates

_genesisMaxLovelaceSupply :: Lovelaces

The total number of lovelace in the system

_genesisNetworkMagic :: Integer

Network identifier

_genesisEpochLength :: Integer

Number of slots in an epoch

_genesisSystemStart :: POSIXTime

Time of slot 0 in UNIX time

_genesisSlotsPerKesPeriod :: Integer

Number of slots in an KES period

_genesisSlotLength :: Integer

Duration of one slot in seconds

_genesisMaxKesEvolutions :: Integer

The maximum number of time a KES key can be evolved before a pool operator must create a new operational certificate

_genesisSecurityParam :: Integer

Security parameter k