Blockfrost.Types.Cardano.Transactions

Description

Cardano Transactions responses

Documentation

data Transaction

Information about a transaction

Constructors

Transaction

Fields

_transactionHash :: Text

Transaction hash

_transactionBlock :: BlockHash

Block hash

_transactionBlockHeight :: Integer

Block number

_transactionSlot :: Slot

Slot number

_transactionIndex :: Integer

Transaction index within the block

_transactionOutputAmount :: [Amount]

Transaction outputs

_transactionFees :: Lovelaces

Fees of the transaction in Lovelaces

_transactionDeposit :: Lovelaces

Deposit within the transaction in Lovelaces

_transactionSize :: Integer

Size of the transaction in Bytes

_transactionInvalidBefore :: Maybe Text

Left (included) endpoint of the timelock validity intervals

_transactionInvalidHereafter :: Maybe Text

Right (excluded) endpoint of the timelock validity intervals

_transactionUtxoCount :: Integer

Count of UTXOs within the transaction

_transactionWithdrawalCount :: Integer

Count of the withdrawals within the transaction

_transactionMirCertCount :: Integer

Count of the MIR certificates within the transaction

_transactionDelegationCount :: Integer

Count of the delegations within the transaction

_transactionStakeCertCount :: Integer

Count of the stake keys (de)registration and delegation certificates within the transaction

_transactionPoolUpdateCount :: Integer

Count of the stake pool registration and update certificates within the transaction

_transactionPoolRetireCount :: Integer

Count of the stake pool retirement certificates within the transaction

_transactionAssetMintOrBurnCount :: Integer

Count of asset mints and burns within the transaction

_transactionRedeemerCount :: Integer

Count of redeemers within the transaction

_transactionValidContract :: Bool

True if this is a valid transaction, False in case of contract validation failure

data TransactionUtxos

Transaction UTxOs

Constructors

TransactionUtxos

Fields

_transactionUtxosHash :: TxHash

Transaction hash

_transactionUtxosInputs :: [UtxoInput]

Transaction inputs

_transactionUtxosOutputs :: [UtxoOutput]

Transaction outputs

data UtxoInput

Transaction input UTxO

Constructors

UtxoInput

Fields

_utxoInputAddress :: Address

Input address

_utxoInputAmount :: [Amount]

_utxoInputTxHash :: TxHash

Hash of the UTXO transaction

_utxoInputOutputIndex :: Integer

UTXO index in the transaction

_utxoInputCollateral :: Bool

UTXO is a script collateral input

_utxoInputDataHash :: Maybe DatumHash

The hash of the transaction output datum

data UtxoOutput

Transaction output UTxO

Constructors

UtxoOutput

Fields

_utxoOutputAddress :: Address

Output address

_utxoOutputAmount :: [Amount]

Transaction output amounts

_utxoOutputDataHash :: Maybe DatumHash

The hash of the transaction output datum

_utxoOutputOutputIndex :: Integer

UTXO index in the transaction

data ValidationPurpose

Validation purpose

Constructors

Spend

Mint

Cert

Reward

data TransactionRedeemer

Transaction redeemer

Constructors

TransactionRedeemer

Fields

_transactionRedeemerTxIndex :: Integer

Index of the redeemer within a transaction

_transactionRedeemerPurpose :: ValidationPurpose

Validation purpose

_transactionRedeemerScriptHash :: ScriptHash

Script hash

_transactionRedeemerDatumHash :: DatumHash

Datum hash

_transactionRedeemerUnitMem :: Quantity

The budget in Memory to run a script

_transactionRedeemerUnitSteps :: Quantity

The budget in Steps to run a script

_transactionRedeemerFee :: Lovelaces

The fee consumed to run the script

data TransactionStake

Information about (de-)registration of a stake address within a transaction

Constructors

TransactionStake

Fields

_transactionStakeCertIndex :: Integer

Index of the certificate within the transaction

_transactionStakeAddress :: Address

Delegation stake address

_transactionStakeRegistration :: Bool

Registration boolean, false if deregistration

data TransactionDelegation

Information about delegation certificates of a specific transaction

Constructors

TransactionDelegation

Fields

_transactionDelegationCertIndex :: Integer

Index of the certificate within the transaction

_transactionDelegationAddress :: Address

Delegation stake address

_transactionDelegationPoolId :: PoolId

Bech32 ID of delegated stake pool

_transactionDelegationActiveEpoch :: Epoch

Epoch in which the delegation becomes active

data TransactionWithdrawal

Information about withdrawals of a specific transaction

Constructors

TransactionWithdrawal

Fields

_transactionWithdrawalAddress :: Address

Bech32 withdrawal address

_transactionWithdrawalAmount :: Lovelaces

Withdrawal amount in Lovelaces

data Pot

Pot from which MIRs are transferred

Constructors

Reserve

Treasury

data TransactionMir

Information about Move Instantaneous Rewards (MIRs) of a specific transaction

Constructors

TransactionMir

Fields

_transactionMirPot :: Pot

Source of MIR funds

_transactionMirCertIndex :: Integer

Index of the certificate within the transaction

_transactionMirAddress :: Address

Bech32 stake address

_transactionMirAmount :: Lovelaces

MIR amount in Lovelaces

data TransactionPoolUpdate

Information about stake pool registration and update certificates of a specific transaction

Constructors

TransactionPoolUpdate

Fields

_transactionPoolUpdateCertIndex :: Integer

Index of the certificate within the transaction

_transactionPoolUpdatePoolId :: PoolId

Bech32 encoded pool ID

_transactionPoolUpdateVrfKey :: Text

VRF key hash

_transactionPoolUpdatePledge :: Lovelaces

Stake pool certificate pledge in Lovelaces

_transactionPoolUpdateMarginCost :: Double

Margin tax cost of the stake pool

_transactionPoolUpdateFixedCost :: Lovelaces

Fixed tax cost of the stake pool in Lovelaces

_transactionPoolUpdateRewardAccount :: Address

Bech32 reward account of the stake pool

_transactionPoolUpdateOwners :: [Address]

_transactionPoolUpdateMetadata :: Maybe PoolUpdateMetadata

_transactionPoolUpdateRelays :: [PoolRelay]

_transactionPoolUpdateActiveEpoch :: Epoch

Epoch that the delegation becomes active

data PoolUpdateMetadata

Update of a pool metadata

Constructors

PoolUpdateMetadata

Fields

_poolUpdateMetadataUrl :: Maybe Text

URL to the stake pool metadata

_poolUpdateMetadataHash :: Maybe Text

Hash of the metadata file

_poolUpdateMetadataTicker :: Maybe Text

Ticker of the stake pool

_poolUpdateMetadataName :: Maybe Text

Name of the stake pool

_poolUpdateMetadataDescription :: Maybe Text

Description of the stake pool

_poolUpdateMetadataHomepage :: Maybe Text

Home page of the stake pool

data TransactionPoolRetiring

Information about stake pool retirements within a specific transaction

Constructors

TransactionPoolRetiring

Fields

_transactionPoolRetiringCertIndex :: Integer

Index of the certificate within the transaction

_transactionPoolRetiringPoolId :: PoolId

Bech32 stake pool ID

_transactionPoolRetiringRetiringEpoch :: Epoch

Retiring epoch

data TransactionMetaJSON

Transaction metadata in JSON

Constructors

TransactionMetaJSON

Fields

_transactionMetaJSONLabel :: Text

Metadata label

_transactionMetaJSONJSONMetadata :: Maybe Value

Content of the JSON metadata

data TransactionMetaCBOR

Transaction metadata in CBOR

Constructors

TransactionMetaCBOR

Fields

_transactionMetaCBORLabel :: Text

Metadata label

_transactionMetaCBORMetadata :: Maybe Text

Content of the CBOR metadata