Blockfrost.Types.IPFS

Description

Types for IPFS servics

Documentation

data IPFSAdd

IPFS Add response

Constructors

IPFSAdd

Fields

_ipfsAddName :: Text

Name of the file

_ipfsAddIpfsHash :: Text

IPFS hash of the file

_ipfsAddSize :: Quantity

Size of the IPFS node

data IPFSPinChange

IPFS Pin Add response

Constructors

IPFSPinChange

Fields

_ipfsPinChangeIpfsHash :: Text

IPFS hash of pinned object

_ipfsPinChangeState :: PinState

State of the pin action

data IPFSPin

IPFS Pin information

Constructors

IPFSPin

Fields

_ipfsPinTimeCreated :: POSIXTime

Creation time of the IPFS object on our backends

_ipfsPinTimePinned :: POSIXTime

Pin time of the IPFS object on our backends

_ipfsPinIpfsHash :: Text

IPFS hash of the pinned object

_ipfsPinSize :: Quantity

Size of the IPFS node

_ipfsPinState :: PinState

State of the pinned object

data PinState

State of the pinned object, which is Queued when we are retriving object.

If this is successful the state is changed to Pinned or Failed if not. The state Gc means the pinned item has been garbage collected due to account being over storage quota or after it has been moved to Unpinned state by removing the object pin.

Constructors

Queued

Pinned

Unpinned

Failed

Gc

newtype IPFSData

Constructors

IPFSData ByteString